added new test
