+++ tests/test_remote.py
@@ -1,5 +1,4 @@
 import logging
-import os
 from unittest.mock import Mock
 
 import pytest
@@ -7,8 +6,6 @@ import pytest
 from block_cascade import GcpEnvironmentConfig, GcpMachineConfig, GcpResource
 from block_cascade import remote
 from block_cascade.utils import PREFECT_VERSION
-from tests.test_config import gcp_resource
-
 if PREFECT_VERSION == 2:
     from prefect.context import FlowRunContext, TaskRunContext
 
@@ -40,24 +37,6 @@ def test_remote():
     """Test the remote decorator with the local executor."""
 
     @remote
-    def addition(a: int, b: int) -> int:
-        return a + b
-
-    result = addition(1, 2)
-
-    assert result == 3
-
-
-def test_remote_local_override():
-    """Test the remote decorator with the local executor using the no_resource_on_local flag."""
-    machine_config = GcpMachineConfig(type="n1-standard-4")
-    gcp_resource = GcpResource(
-        chief=machine_config,
-        environment=GcpEnvironmentConfig(
-            storage_location=GCP_STORAGE_LOCATION, project=GCP_PROJECT
-        ),
-    )
-    @remote(resource=gcp_resource, no_resource_on_local=True)
     def addition(a: int, b: int) -> int:
         return a + b
